🔨 Refactor Slack message handling into dedicated helper functions
